Moved tets to their own file
